% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_fancy.R
\name{theme_fancy}
\alias{theme_fancy}
\title{Adding a color to the background.}
\usage{
theme_fancy(fill_color = "#ededed")
}
\arguments{
\item{fill_color, }{background color}
}
\description{
Adding a color to the background.
}
