% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_raleway.R
\name{theme_raleway}
\alias{theme_raleway}
\title{Minimal ggplot2 theme using Raleway fonts.
Functionally a Raleway spin on (https://github.com/hrbrmstr/hrbrthemes/)}
\usage{
theme_raleway(base_family = "Raleway", base_size = 11.5,
  plot_title_family = base_family, plot_title_size = 16,
  plot_title_margin = 10, plot_title_face = "plain",
  subtitle_family = "Raleway", subtitle_size = 13,
  subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", caption_family = "Raleway",
  caption_size = 10, caption_face = "plain", caption_margin = 10,
  caption = NULL, axis_text_size = base_size,
  axis_title_family = base_family, axis_title_size = 9,
  axis_title_face = "plain", axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30), grid_col = "#cccccc", grid = TRUE,
  axis_col = "#cccccc", axis = FALSE, ticks = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid_col}{grid color}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\description{
Minimal ggplot2 theme using Raleway fonts.
Functionally a Raleway spin on (https://github.com/hrbrmstr/hrbrthemes/)
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg), col = displ) +
  geom_point() +
  theme_Raleway() +
  labs(x = "omg this is very much the x axis",
    y = "yes definitely this indeed is the y axis",
    title = "Testing Testing: here's the thing",
    subtitle = "I am a subtitle so I am smaller",
    caption = label_wrap_gen(120)("wow when do you actually ever include captions in your ggplots??? just throw it in your Fig. 1 description and keep it moving. but fine, this is what it would look like if you insist.")
    )
}
}
